docs(payouts): add comprehensive payout algorithm documentation
## Documentation (docs/payouts/algorithm/)
- 00-readme.md: Overview, navigation, quick start checklist
- 01-architecture.md: System diagrams (mermaid), service layers, ER diagram
- 02-earnings-lifecycle.md: Status flow, hold periods, revenue split (80/20)
- 03-payout-processing.md: Batch creation, approval workflow, webhooks
- 04-api-reference.md: Complete endpoint documentation
- 05-configuration.md: Environment variables, constants, provider setup

## Refactor: Extract shared dashboard components
- Create components/dashboard/shared/InvoicesPage.tsx
- Reduce admin/staff invoices pages from 291 → 15 lines each
- Remove ~550 lines of duplicated code
