+++ components/dashboard/shared/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Shared Dashboard Components
+ * Reusable components for admin and staff dashboards
+ */
+
+export { InvoicesPage } from "./InvoicesPage";
